+++ pkg/tsdb/prometheus/models/query.go
@@ -14,8 +14,59 @@ import (
 	"github.com/prometheus/prometheus/promql/parser"
 
 	"github.com/grafana/grafana/pkg/tsdb/prometheus/intervalv2"
-	"github.com/grafana/grafana/pkg/tsdb/prometheus/kinds"
 )
+
+// PromQueryFormat defines model for PromQueryFormat.
+// +enum
+type PromQueryFormat string
+
+const (
+	PromQueryFormatTimeSeries PromQueryFormat = "time_series"
+	PromQueryFormatTable      PromQueryFormat = "table"
+	PromQueryFormatHeatmap    PromQueryFormat = "heatmap"
+)
+
+// QueryEditorMode defines model for QueryEditorMode.
+// +enum
+type QueryEditorMode string
+
+const (
+	QueryEditorModeBuilder QueryEditorMode = "builder"
+	QueryEditorModeCode    QueryEditorMode = "code"
+)
+
+// PrometheuQueryProperties defines the specific properties used for prometheus
+type PrometheuQueryProperties struct {
+	// The response format
+	Format PromQueryFormat `json:"format,omitempty"`
+
+	// The actual expression/query that will be evaluated by Prometheus
+	Expr string `json:"expr"`
+
+	// Returns a Range vector, comprised of a set of time series containing a range of data points over time for each time series
+	Range bool `json:"range,omitempty"`
+
+	// Returns only the latest value that Prometheus has scraped for the requested time series
+	Instant bool `json:"instant,omitempty"`
+
+	// Execute an additional query to identify interesting raw samples relevant for the given expr
+	Exemplar bool `json:"exemplar,omitempty"`
+
+	// what we should show in the editor
+	EditorMode QueryEditorMode `json:"editorMode,omitempty"`
+
+	// @deprecated Used to specify how many times to divide max data points by. We use max data points under query options
+	// See https://github.com/grafana/grafana/issues/48081
+	IntervalFactor *float32 `json:"intervalFactor,omitempty"`
+
+	// Series name override or template. Ex. {{hostname}} will be replaced with label value for hostname
+	LegendFormat string `json:"legendFormat,omitempty"`
+
+	// ???
+	Scope *struct {
+		Matchers string `json:"matchers"`
+	} `json:"scope,omitempty"`
+}
 
 // Internal interval and range variables
 const (
@@ -51,16 +102,23 @@ const (
 
 var safeResolution = 11000
 
+// QueryModel includes both the common and specific values
 type QueryModel struct {
-	kinds.PrometheusDataQuery
-	RefId string `json:"refId,omitempty"`
+	PrometheuQueryProperties `json:",inline"`
+	CommonQueryProperties    `json:",inline"`
 
 	// The following properties may be part of the request payload, however they are not saved in panel JSON
 	// Timezone offset to align start & end time on backend
 	UtcOffsetSec   int64  `json:"utcOffsetSec,omitempty"`
 	Interval       string `json:"interval,omitempty"`
-	IntervalMs     int64  `json:"intervalMs,omitempty"`
 	IntervalFactor int64  `json:"intervalFactor,omitempty"`
+}
+
+// CommonQueryProperties is properties applied to all queries
+// NOTE: this will soon be replaced with a struct from the SDK
+type CommonQueryProperties struct {
+	RefId      string `json:"refId,omitempty"`
+	IntervalMs int64  `json:"intervalMs,omitempty"`
 }
 
 type TimeRange struct {
